reconfigure project structure
