+++ examples/main.cpp
@@ -1,16 +1,17 @@
-#include "../include/LongArithmetic.hpp"
+#include "../include/LongNumMod.hpp"
 
 int main() {
-    LongNumber num1("E775FA3C0F3E1CDD13C963C64166D0DE107E98E1523DD70C9");
-    LongNumber num2("E775FA3C0F3E1CDD13C963C64166D0DE10719246781672487");
-    LongNumber modulo("E1CDD13C963C64166D0DE107E98E1523DD70C9E3AE314");
+    LongNumber num1("104D4A301BD70480050C46201F6AA95018228D200F5CA420252");
+    LongNumber num2("8");
+    LongNumber modulus("104D4A301BD70480050C46201F6AA95018228D200F5CA420252104D4A301BD70480050C46201F6AA95018228D200F5CA420252104D4A301BD70480050C46201F6AA95018228D200F5CA420252104D4A301BD70480050C46201F6AA95018228D200F5CA420252104D4A301BD70480050C46201F6AA95018228D200F5CA420252104D4A301BD70480050C46201F6AA95018228D200F5CA420252104D4A301BD70480050C46201F6AA95018228D200F5CA420252104D4A301BD70480050C46201F6AA95018228D200F5CA420252");
 
-    LongNumberMod ln1(num1, modulo);
-    LongNumberMod ln2(num2, modulo);
+    LongNumberMod ln1(num1);
+    const LongNumberMod ln2(num2);
+    ln1.initializeModulus(modulus);
 
-    ln1 + ln2;
-    ln1 - ln2;
-    
+    LongNumberMod ans = ln1.toPowerOf(ln2);
+
+    std::cout << ans.getNumHex() << std::endl;    
 
     // std::cout << num1.toHexString() << std::endl;
     // std::cout << num2.toHexString() << std::endl;
