+++ measurment/averageTime.cpp
@@ -1,4 +1,4 @@
-#include "../include/LongArithmetic.hpp"
+#include "../include/LongNumMod.hpp"
 #include <chrono>
 #include <thread>
