+++ tests/CMakeLists.txt
@@ -1,13 +1,14 @@
 find_package(GTest REQUIRED)
 
-add_executable(LongArithmeticLibTests
-    LongArithmeticTests.cpp
+add_executable(LongArithmeticTests
+    LongNumTests.cpp
+    LongNumModTests.cpp
 )
 
-target_link_libraries(LongArithmeticLibTests
+target_link_libraries(LongArithmeticTests
     LongArithmeticLib
     GTest::Main  
 )
 
 include(GoogleTest)
-gtest_discover_tests(LongArithmeticLibTests)
+gtest_discover_tests(LongArithmeticTests)
